Remove Button version 1. Replace 1 by version 2 and renamed as the official button component. Remove all version props
