+++ src/components/AvatarStack/AvatarStackV2.tsx
@@ -3,7 +3,7 @@ import AvatarStackBase from './AvatarStack'
 
 export class AvatarStack extends React.PureComponent {
   render() {
-    return <AvatarStackBase {...this.props} avatarVersion={2} version={2} />
+    return <AvatarStackBase {...this.props} avatarVersion={2} />
   }
 }
